+++ Examples/TicTacToe/Sources/RootView.swift
@@ -1,5 +1,7 @@
 import AppCore
 import AppSwiftUI
+import AppUIKit
+import AuthenticationClientLive
 import ComposableArchitecture
 import SwiftUI
