--- Examples/TicTacToe/Sources/SceneDelegate.swift
@@ -1,30 +1,0 @@
-import AppCore
-import AppSwiftUI
-import ComposableArchitecture
-import LiveAuthenticationClient
-import SwiftUI
-import UIKit
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-  var window: UIWindow?
-
-  func scene(
-    _ scene: UIScene,
-    willConnectTo session: UISceneSession,
-    options connectionOptions: UIScene.ConnectionOptions
-  ) {
-    self.window = (scene as? UIWindowScene).map(UIWindow.init(windowScene:))
-    self.window?.rootViewController = UIHostingController(rootView: RootView())
-    self.window?.makeKeyAndVisible()
-  }
-}
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-  func application(
-    _ application: UIApplication,
-    didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
-  ) -> Bool {
-    true
-  }
-}
