+++ Examples/TicTacToe/Sources/TicTacToeApp.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+@main
+struct TicTacToeApp: App {
+  var body: some Scene {
+    WindowGroup {
+      RootView()
+    }
+  }
+}
