+++ Examples/TicTacToe/tic-tac-toe/Sources/AppSwiftUI/AppView.swift
@@ -1,0 +1,30 @@
+import AppCore
+import ComposableArchitecture
+import LoginSwiftUI
+import NewGameSwiftUI
+import SwiftUI
+
+public struct AppView: View {
+  let store: Store<AppState, AppAction>
+
+  public init(store: Store<AppState, AppAction>) {
+    self.store = store
+  }
+
+  public var body: some View {
+    SwitchStore(self.store) {
+      CaseLet(state: /AppState.login, action: AppAction.login) { store in
+        NavigationView {
+          LoginView(store: store)
+        }
+        .navigationViewStyle(StackNavigationViewStyle())
+      }
+      CaseLet(state: /AppState.newGame, action: AppAction.newGame) { store in
+        NavigationView {
+          NewGameView(store: store)
+        }
+        .navigationViewStyle(StackNavigationViewStyle())
+      }
+    }
+  }
+}
